+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Unreleased
   - Adjusted various symbolization code paths to stop heap-allocating
 - Adjusted normalization logic to honor executable and readable proc maps
   entries
+- Changed `debug_syms` to be a symbolization source property instead of a
+  `symbolize::Symbolizer` attribute
 - Renamed `inspect::Elf::debug_info` to `debug_syms`
 - Handled potential numeric overflow in Gsym inlined function parser more
   gracefully